refactor(mcp): extract resource and prompt catalog
